A bunch of uint64 -> int64

+++ freespace_unix.go
@@ -6,8 +6,8 @@ import (
 	"syscall"
 )
 
-func filesystemFree() (uint64, error) {
+func filesystemFree() (int64, error) {
 	fs := syscall.Statfs_t{}
 	err := syscall.Statfs(*root, &fs)
-	return fs.Bfree * uint64(fs.Bsize), err
+	return int64(fs.Bfree) * int64(fs.Bsize), err
 }
